Author: Francesco Chemolli <kinkie@squid-cache.org> String NG : basic SquidString updates pt 2
  - finished the analysis of raw buffer exporting, divided among c-sting
  and raw-buf access clients
  - general sanitization
  - reversal of some wrongly-renamed variables and comments
  - implementation of SQUIDSTRINGPRINT and SQUIDSTRINGPH (placeholder) macros and
      psize() function for printf-style calls
  - implementation of, and migration to, String::size_type and String::npos
  - de-inlining of pos(), rpos(), find() and rfind() calls
  - implementation and use of a proper substr() call

+++ src/HttpHeader.cc
@@ -1683,7 +1683,7 @@ httpHeaderStoreReport(StoreEntry * e)
 }
 
 http_hdr_type
-httpHeaderIdByName(const char *name, int name_len, const HttpHeaderFieldInfo * info, int end)
+httpHeaderIdByName(const char *name, size_t name_len, const HttpHeaderFieldInfo * info, int end)
 {
     int i;
